AWLLoginError now a fata exception
Causes server startup to fail if credentials are incorrect.
Hypercorn will fall into a permanently failed state if
AWL server credentials are changed while it is running,
and it must be restarted with the new credentials configured.
There is currently no fix other than reconfiguring and restarting,
so this permanently failed state is not the worst thing in
the world. The better solution would be for Quart to hook
into `shutdown_trigger` on hypercorn's `serve()` function,
but that requires upstream changes to Quart.